add card reader
The card reader listens to events from an RFID card reader. The RFID
card reader is found from the list of input devices with a fixed
string.

If an ENTER key is detected the reader has finished reading an RFID
card. In this case the read card ID is logged and the reader is reset
to start listening for the next card.