manage playlist playback via chromecast
Since we cannot really control google music on the chromecast device,
we need to manage the playlist playback manually. The app will
register for chromecast events to queue the next song if playback of
the current song finishes.

Playbacks stops if the chromecast playback was interrupted by other
sources, e.g. from the phone.